Switch from single hold to multi-hold
Instead of pressing the `HoldKey` and waiting for the next key press
now the plugin records all held keys when the `HoldKey` is released,
and continues to hold them down. By default the plugin will hold up
to 6 different keys (configurable with #define). Pressing any key
still cancels the holding of keys.